refactor: move `delta_time` to `f64` & feat: atomic frame counting (#39)
## Changes

- add: `frame` in engine. An atomic u64 for frame count
- add: `BeginFrame` event to broadcast the frame count at the beginning
of every frame.
- refactor: move all delta times to use by `f64`

+++ crates/dirk_renderer/src/lib.rs
@@ -494,7 +494,7 @@ impl Renderer {
     /// world (unless in [`WorldEvent::Created`] or [`WorldEvent::Destroyed`].
     pub fn tick(
         &mut self,
-        _delta_time: f32,
+        _delta_time: f64,
         windows: &HashMap<WindowId, dirk_platform::Window>,
     ) -> Result<()> {
         // Temporarily move receivers out for the borrow checker
